requirements.txt
python-telegram-bot==20.7
Pillow==10.1.0
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
